requirements_api.txt
fastapi
mongoengine
numpy
pillow
